Introduce managed properties in resources
Ok, so this is where it starts to get out of hand... I'm adding a sort of
prototype class that defines validation parameters and the DSL definitions
for scalar and list properties. These are defined at the Class-level.

Additionally, there's a porperty-instance class that gets instiantated with
the parent resource. This class defines the getter/setter interface for the
property, and uses the validators/transforms defined in the prototype.

Several of the DSL helpers now leverage the validator/transform interface.