Maintain backwards compatability with route53
Previously, users needed to specify the json for alias_target and
geo_location for a route53_recordset.  In the add-alb branch, merged in
PR 214, this was changed in a backwards-incompatible way, requiring
anyone who had previously specified json to now specify a block for the
resource_property.  Now, either way is acceptable.

This is potentially a paradigm that can be used elsewhere, as there are
many places where a resource_property would be appropriate, but has not
existed and users have implemented the functionality directly with a
json (or ruby hash) object.  This will allow us to deprecate that method
in favor of the more clean resource_property approach, without breaking
everything in a 1.0 upgrade.