Fix diff for Mappings
The Mapings section of CloudFormation templates is treated the same way
by the diff flattener as the Resources section.  Unfortunately, when
updating the diff to be smart enough to know about updates which require
replacement, the Resource Type was added to the flattened key
unconditionally.  Since Mappings (and other not-Resources) have no
"Type" key, this meant an empty hash was added where the Type belonged.
To fix this, we now just don't put the Type in the flattened hash key
if there isn't one.