Implement API Gateway resources and properties
There are a lot here and I haven't tested all of them but the ones I
have seemed to work.  Also added the Fn::Sub intrinsic function because
it's used by amazon in one of the examples for how api gateway can be
implemented, and I tested that the non-hash version works.