Make info messages cyan instead of green
When running a `diff --vv`, it's sometimes hard to see one change in a
sea of unchanged stack info messages.  This is because the `compare` and
`create` and `update` look very similar given the color.  With this new
change, `compare` and `unchanged` will be cyan, but `create` and
`update` will continue to show as green.

This has been tested on both a black and a white background, so should
have minimal annoyance factor across a range of color palates.